+++ packages/melonjs/src/video/webgl/buffer/vertex.js
@@ -1,22 +1,21 @@
 /**
- * a Vertex Buffer object
+ * a fixed-size Vertex Buffer object.
+ * Compositors must check isFull() and flush before the buffer overflows.
  * @ignore
  */
 
 export default class VertexArrayBuffer {
-	constructor(vertex_size, vertex_per_obj) {
-		// the size of one vertex in float
-		this.vertexSize = vertex_size;
-		// size of an object in vertex
-		this.objSize = vertex_per_obj;
+	constructor(vertexSize, maxVertex) {
+		// the size of one vertex in floats
+		this.vertexSize = vertexSize;
 		// the maximum number of vertices the vertex array buffer can hold
-		this.maxVertex = 256; // (note: this seems to be the sweet spot performance-wise when using batching)
+		this.maxVertex = maxVertex;
 		// the current number of vertices added to the vertex array buffer
 		this.vertexCount = 0;
 
 		// the actual vertex data buffer
 		this.buffer = new ArrayBuffer(
-			this.maxVertex * this.vertexSize * this.objSize,
+			this.maxVertex * this.vertexSize * Float32Array.BYTES_PER_ELEMENT,
 		);
 		// Float32 and Uint32 view of the vertex data array buffer
 		this.bufferF32 = new Float32Array(this.buffer);
@@ -35,34 +34,8 @@ export default class VertexArrayBuffer {
 	 * return true if full
 	 * @ignore
 	 */
-	isFull(vertex = this.objSize) {
+	isFull(vertex) {
 		return this.vertexCount + vertex >= this.maxVertex;
-	}
-
-	/**
-	 * resize the vertex buffer, retaining its original contents
-	 * @ignore
-	 */
-	resize(vertexCount) {
-		while (vertexCount > this.maxVertex) {
-			// double the vertex size
-			this.maxVertex <<= 1;
-		}
-
-		// save a reference to the previous data
-		const data = this.bufferF32;
-
-		// recreate ArrayBuffer and views
-		this.buffer = new ArrayBuffer(
-			this.maxVertex * this.vertexSize * this.objSize,
-		);
-		this.bufferF32 = new Float32Array(this.buffer);
-		this.bufferU32 = new Uint32Array(this.buffer);
-
-		// copy previous data
-		this.bufferF32.set(data);
-
-		return this;
 	}
 
 	/**
@@ -72,21 +45,11 @@ export default class VertexArrayBuffer {
 	push(x, y, u, v, tint) {
 		let offset = this.vertexCount * this.vertexSize;
 
-		if (this.vertexCount >= this.maxVertex) {
-			this.resize(this.vertexCount);
-		}
-
 		this.bufferF32[offset] = x;
 		this.bufferF32[++offset] = y;
-
-		if (typeof u !== "undefined") {
-			this.bufferF32[++offset] = u;
-			this.bufferF32[++offset] = v;
-		}
-
-		if (typeof tint !== "undefined") {
-			this.bufferU32[++offset] = tint;
-		}
+		this.bufferF32[++offset] = u;
+		this.bufferF32[++offset] = v;
+		this.bufferU32[++offset] = tint;
 
 		this.vertexCount++;
 
@@ -115,21 +78,5 @@ export default class VertexArrayBuffer {
 		} else {
 			return this.bufferU32;
 		}
-	}
-
-	/**
-	 * return the size of the vertex in vertex
-	 * @ignore
-	 */
-	length() {
-		return this.vertexCount;
-	}
-
-	/**
-	 * return true if empty
-	 * @ignore
-	 */
-	isEmpty() {
-		return this.vertexCount === 0;
 	}
 }